services/ml_service/requirements.txt
# ML Service requirements
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
openai>=1.10.0
numpy>=1.26.0
scikit-learn>=1.4.0
kafka-python>=2.0.2
prometheus-client>=0.19.0
tenacity>=8.2.0

